tasks: legacy tasks and workflow removal
* INCOMPATIBLE Removes legacy task functions used in workflows
  that are deprecated along with legacy workflow definitions.

* NEW Adds new workflows and tasks that is meant to work with
  Invenio Records.
