--- invenio_oaiharvester/workflows/oaiharvest_repository_with_approval.py
@@ -1,29 +1,0 @@
-# -*- coding: utf-8 -*-
-#
-# This file is part of Invenio.
-# Copyright (C) 2014, 2015 CERN.
-#
-# Invenio is free software; you can redistribute it and/or
-# modify it under the terms of the GNU General Public License as
-# published by the Free Software Foundation; either version 2 of the
-# License, or (at your option) any later version.
-#
-# Invenio is distributed in the hope that it will be useful, but
-# WITHOUT ANY WARRANTY; without even the implied warranty of
-# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the GNU
-# General Public License for more details.
-#
-# You should have received a copy of the GNU General Public License
-# along with Invenio; if not, write to the Free Software Foundation, Inc.,
-# 59 Temple Place, Suite 330, Boston, MA 02111 1307, USA.
-
-"""Main workflow iterating over selected repositories and downloaded files."""
-
-from .oaiharvest_harvest_repositories import oaiharvest_harvest_repositories
-
-
-class oaiharvest_repository_with_approval(oaiharvest_harvest_repositories):
-
-    """A workflow for use with OAI harvesting in BibSched with approval."""
-
-    record_workflow = "oaiharvest_record_approval"
